+++ marsworks/__init__.py
@@ -22,8 +22,9 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-__version__ = "0.4.2"
+__version__ = "0.5.0"
 from marsworks.client import *
 from marsworks.manifest import *
 from marsworks.origin import *
 from marsworks.photo import *
+from marsworks.partialmanifest import *