+++ marsworks/origin/serializer.py
@@ -1,4 +1,27 @@
-import inspect
+"""
+MIT License
+
+Copyright (c) 2021 NovaEmiya
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
+
 from typing import Optional
 
 import httpx
@@ -6,6 +29,7 @@ import marsworks
 
 from ..manifest import Manifest
 from .exceptions import BadContentError
+from .internal_utils import repr_gen
 
 __all__ = ("Serializer",)
 
@@ -61,17 +85,10 @@ class Serializer:
 
         raise BadContentError(content=data)
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         """
         Returns:
 
             Representation of Serializer. (Result of `repr(obj)`)
         """
-        attrs = [
-            attr
-            for attr in inspect.getmembers(self)
-            if not inspect.ismethod(attr[1])
-            if not attr[0].startswith("_")
-        ]
-        fmt = ", ".join(f"{attr}={value}" for attr, value in attrs)[:-2]
-        return f"{__class__.__name__}({fmt})"
+        return repr_gen(__class__, self)